+++ backend/interfaces/api/server/handler/article.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 
 	"github.com/gorilla/mux"
-	"github.com/yoshihiro-shu/draft-backend/auth"
+	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server/auth"
 	"github.com/yoshihiro-shu/draft-backend/model"
 	article_linkages_to_category "github.com/yoshihiro-shu/draft-backend/model/article/linkages/to/category"
 	article_linkages_to_many "github.com/yoshihiro-shu/draft-backend/model/article/linkages/to/many"
