+++ backend/interfaces/api/server/router/setRouters.go
@@ -3,8 +3,8 @@ package router
 import (
 	"net/http"
 
-	"github.com/yoshihiro-shu/draft-backend/auth"
 	"github.com/yoshihiro-shu/draft-backend/controllers/api"
+	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server/auth"
 	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server/handler"
 	"github.com/yoshihiro-shu/draft-backend/request"
 )
