+++ backend/request/context.go
@@ -9,9 +9,9 @@ import (
 	"os"
 
 	"github.com/go-pg/pg"
-	"github.com/yoshihiro-shu/draft-backend/auth"
 	"github.com/yoshihiro-shu/draft-backend/cache"
 	"github.com/yoshihiro-shu/draft-backend/config"
+	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server/auth"
 	"github.com/yoshihiro-shu/draft-backend/model"
 )
